requirements.txt
# ── Core ──────────────────────────────────────────────────────────────
python-dotenv
requests

# ── Data & Database ──────────────────────────────────────────────────
sqlalchemy
pandas
numpy

# ── Feature Engineering ──────────────────────────────────────────────
ta

# ── Machine Learning ─────────────────────────────────────────────────
scikit-learn
xgboost
joblib

# ── API ──────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── Dashboard ────────────────────────────────────────────────────────
streamlit
plotly

-e .
